Implement conditional expression parsing and utility functions
- Added `Condition` and `ConditionValue` enums to represent logical conditions and their values.
- Implemented expression tokenization and parsing logic in `expressions.rs`.
- Introduced utility functions for handling string processing, including unescaping and parsing stance prefixes in `utils.rs`.
- Enhanced tests to cover new functionality, including fallback strings and verb conjugation scenarios.
