+++ src/engine.rs
@@ -1,8 +1,8 @@
 use crate::grammar::{
     capitalize_cow, conjugate_verb, push_capitalized_if, resolve_article, resolve_pronoun,
 };
-use crate::models::{NULL_VIEWER, RenderContext, TemplateEntity, is_same_entity};
-use crate::parser::{MOD_POSSESSIVE, TAG_PROPERTY_SEP};
+use crate::models::{NULL_VIEWER, RenderContext, TemplateEntity};
+use crate::parser::MOD_POSSESSIVE;
 pub use crate::parser::{TagFlags, Template, Token};
 use crate::typography::{
     NO_SENTENCE_BREAK_SENTINEL, SENTENCE_BREAK_SENTINEL, apply_all_caps, post_process_typography,
@@ -11,8 +11,15 @@ use crate::typography::{
 use crate::typography::{has_protocol_tags, skip_protocol_tags};
 use std::collections::HashMap;
 
+/// Anaphora resolution and memory mutation logic.
+pub mod anaphora;
+/// Disambiguation and naming collision logic.
+pub mod disambiguation;
+/// Group formatting, Oxford comma distribution, and nested list flattening logic.
+pub mod groups;
+
 /// Parameters extracted from a token or fallback logic to render an entity.
-struct EntityRefParams<'p> {
+pub(crate) struct EntityRefParams<'p> {
     key: &'p str,
     article: Option<&'p str>,
     p_type: Option<&'p str>,
@@ -34,15 +41,7 @@ enum PronounResult<'a> {
 
 bitflags::bitflags! {
     #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-    struct GroupMemberFlags: u8 {
-        const AFTER_POSSESSIVE       = 1 << 0;
-        const FIRST_VISIBLE_ITEM     = 1 << 1;
-        const DISTRIBUTE_POSSESSIVES = 1 << 2;
-        const IS_REFLEXIVE           = 1 << 3;
-    }
-
-    #[derive(Clone, Copy, Debug, PartialEq, Eq)]
-    struct PronounContextFlags: u8 {
+    pub(crate) struct PronounContextFlags: u8 {
         const IS_PLURAL         = 1 << 0;
         const IS_GROUP          = 1 << 1;
         const IS_VIEWER         = 1 << 2;
@@ -58,12 +57,6 @@ bitflags::bitflags! {
         const IS_CAPITALIZED = 1 << 2;
         const IS_OBJ         = 1 << 3;
     }
-}
-
-/// Configuration for formatting a group member.
-struct GroupMemberFormatConfig<'a> {
-    flags: GroupMemberFlags,
-    article_to_use: Option<&'a str>,
 }
 
 /// The core processor responsible for evaluating compiled templates against contexts.
@@ -101,7 +94,7 @@ impl PerspectiveEngine {
         // "Other") and long description disambiguation.
         let mut pre_resolved: HashMap<&str, &dyn TemplateEntity> = HashMap::new();
         for k in &template.template_keys {
-            if let Ok(ent) = Self::get_entity(ctx, k) {
+            if let Ok(ent) = crate::evaluator::get_entity(ctx, k) {
                 pre_resolved.insert(k.as_str(), ent);
             }
         }
@@ -208,22 +201,22 @@ impl PerspectiveEngine {
             return Ok(());
         };
 
-        let resolved_key = Self::resolve_tag_segment(ctx, key, pre_resolved)?;
+        let resolved_key = crate::evaluator::resolve_tag_segment(ctx, key, pre_resolved)?;
         let resolved_article = article
             .as_ref()
-            .map(|a| Self::resolve_tag_segment(ctx, a, pre_resolved))
+            .map(|a| crate::evaluator::resolve_tag_segment(ctx, a, pre_resolved))
             .transpose()?;
         let resolved_p_type = p_type
             .as_ref()
-            .map(|p| Self::resolve_tag_segment(ctx, p, pre_resolved))
+            .map(|p| crate::evaluator::resolve_tag_segment(ctx, p, pre_resolved))
             .transpose()?;
         let resolved_owner_key = owner_key
             .as_ref()
-            .map(|o| Self::resolve_tag_segment(ctx, o, pre_resolved))
+            .map(|o| crate::evaluator::resolve_tag_segment(ctx, o, pre_resolved))
             .transpose()?;
         let resolved_adjectives = adjectives
             .as_ref()
-            .map(|a| Self::resolve_tag_segment(ctx, a, pre_resolved))
+            .map(|a| crate::evaluator::resolve_tag_segment(ctx, a, pre_resolved))
             .transpose()?;
 
         Self::render_entity_ref(
@@ -258,7 +251,7 @@ impl PerspectiveEngine {
 
         let mut matched = false;
         for branch in branches {
-            if Self::evaluate_condition(ctx, &branch.condition, pre_resolved) {
+            if crate::evaluator::evaluate_condition(ctx, &branch.condition, pre_resolved) {
                 Self::render_tokens(
                     ctx,
                     &branch.body,
@@ -277,568 +270,6 @@ impl PerspectiveEngine {
         Ok(())
     }
 
-    #[inline]
-    fn get_entity<'a>(ctx: &'a RenderContext, key: &str) -> Result<&'a dyn TemplateEntity, String> {
-        // 1. Try exact match first (e.g., "source")
-        if let Some(entity) = ctx.entities.get(key).copied() {
-            return Ok(entity);
-        }
-
-        // 2. Try dot notation traversal (e.g., "source.left_arm.weapon")
-        if let Some((root_key, remainder)) = key.split_once(TAG_PROPERTY_SEP)
-            && let Some(mut current) = ctx.entities.get(root_key).copied()
-        {
-            let mut current_path = root_key;
-            for prop in remainder.split(TAG_PROPERTY_SEP) {
-                current = current.get_property(prop).ok_or_else(|| {
-                    format!("Missing property '{prop}' on entity '{current_path}'")
-                })?;
-                // Accumulate the traversed path slice by extending it over the dot and property
-                let next_len = current_path.len() + 1 + prop.len();
-                current_path = key.get(..next_len).unwrap_or(current_path);
-            }
-            return Ok(current);
-        }
-
-        tracing::error!("Failed to render template: Missing entity for key '{key}'");
-        Err(format!("Missing entity for key: {key}"))
-    }
-
-    #[inline]
-    fn resolve_tag_segment<'a>(
-        ctx: &'a RenderContext,
-        segment: &'a crate::parser::TagSegment,
-        pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
-    ) -> Result<std::borrow::Cow<'a, str>, String> {
-        match segment {
-            crate::parser::TagSegment::Literal(s) => Ok(std::borrow::Cow::Borrowed(s.as_str())),
-            crate::parser::TagSegment::Variable { key: k, fallback } => {
-                if let Some((ent, prop)) = k.rsplit_once('.') {
-                    let root_key = ent.split_once('.').map_or(ent, |(r, _)| r);
-                    if ctx.entities.contains_key(root_key) {
-                        if let Ok(entity) = pre_resolved
-                            .get(ent)
-                            .copied()
-                            .map_or_else(|| Self::get_entity(ctx, ent), Ok)
-                            && let Some(val) = entity.get_string_property(prop)
-                        {
-                            return Ok(val);
-                        }
-                        if let Some(fb) = fallback {
-                            return Ok(std::borrow::Cow::Owned(fb.clone()));
-                        }
-                        tracing::error!("Missing string property '{prop}' on entity '{ent}'");
-                        return Err(format!(
-                            "Missing string property '{prop}' on entity '{ent}'"
-                        ));
-                    }
-                }
-                if let Some(values) = ctx.variables.get(k.as_str()) {
-                    if let Some(v) = values.first() {
-                        Ok(std::borrow::Cow::Borrowed(v.as_str()))
-                    } else {
-                        Ok(std::borrow::Cow::Borrowed(""))
-                    }
-                } else {
-                    if let Some(fb) = fallback {
-                        return Ok(std::borrow::Cow::Owned(fb.clone()));
-                    }
-                    tracing::error!("Missing dynamic variable for key '{k}'");
-                    Err(format!("Missing variable for key: {k}"))
-                }
-            }
-        }
-    }
-
-    fn evaluate_condition_value_bool(
-        ctx: &RenderContext,
-        val: &crate::parser::ConditionValue,
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-    ) -> bool {
-        match val {
-            crate::parser::ConditionValue::Literal(s) => {
-                !s.is_empty() && !s.eq_ignore_ascii_case("false") && s != "0"
-            }
-            crate::parser::ConditionValue::Number(n) => *n != 0.0,
-            crate::parser::ConditionValue::Variable(var) => {
-                ctx.variables.get(var.as_str()).is_some_and(|v| {
-                    v.first().is_some_and(|f| {
-                        !f.is_empty() && !f.eq_ignore_ascii_case("false") && f != "0"
-                    })
-                })
-            }
-            crate::parser::ConditionValue::EntityProperty(ent, prop) => {
-                if let Ok(entity) = pre_resolved
-                    .get(ent.as_str())
-                    .copied()
-                    .map_or_else(|| Self::get_entity(ctx, ent), Ok)
-                {
-                    entity.check_condition(prop)
-                } else {
-                    false
-                }
-            }
-        }
-    }
-
-    fn evaluate_condition_value_string<'a>(
-        ctx: &'a RenderContext,
-        val: &'a crate::parser::ConditionValue,
-        pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
-    ) -> Option<std::borrow::Cow<'a, str>> {
-        match val {
-            crate::parser::ConditionValue::Literal(s) => {
-                Some(std::borrow::Cow::Borrowed(s.as_str()))
-            }
-            crate::parser::ConditionValue::Number(n) => {
-                Some(std::borrow::Cow::Owned(n.to_string()))
-            }
-            crate::parser::ConditionValue::Variable(var) => ctx
-                .variables
-                .get(var.as_str())
-                .and_then(|v| v.first())
-                .map(|s| std::borrow::Cow::Borrowed(s.as_str())),
-            crate::parser::ConditionValue::EntityProperty(ent, prop) => {
-                if let Ok(entity) = pre_resolved
-                    .get(ent.as_str())
-                    .copied()
-                    .map_or_else(|| Self::get_entity(ctx, ent), Ok)
-                {
-                    entity.get_string_property(prop)
-                } else {
-                    None
-                }
-            }
-        }
-    }
-
-    fn get_numeric_value(
-        ctx: &RenderContext,
-        val: &crate::parser::ConditionValue,
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-    ) -> Option<f64> {
-        if let crate::parser::ConditionValue::Number(n) = val {
-            Some(*n)
-        } else {
-            let s_opt = Self::evaluate_condition_value_string(ctx, val, pre_resolved);
-            s_opt.and_then(|s| s.parse::<f64>().ok())
-        }
-    }
-
-    fn evaluate_numeric(
-        ctx: &RenderContext,
-        left: &crate::parser::ConditionValue,
-        right: &crate::parser::ConditionValue,
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-        op: impl Fn(f64, f64) -> bool,
-    ) -> bool {
-        let left_num = Self::get_numeric_value(ctx, left, pre_resolved);
-        let right_num = Self::get_numeric_value(ctx, right, pre_resolved);
-        if let (Some(l), Some(r)) = (left_num, right_num) {
-            op(l, r)
-        } else {
-            false
-        }
-    }
-
-    fn evaluate_condition(
-        ctx: &RenderContext,
-        condition: &crate::parser::Condition,
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-    ) -> bool {
-        match condition {
-            crate::parser::Condition::Value(val) => {
-                Self::evaluate_condition_value_bool(ctx, val, pre_resolved)
-            }
-            crate::parser::Condition::Not(inner) => {
-                !Self::evaluate_condition(ctx, inner, pre_resolved)
-            }
-            crate::parser::Condition::And(left, right) => {
-                Self::evaluate_condition(ctx, left, pre_resolved)
-                    && Self::evaluate_condition(ctx, right, pre_resolved)
-            }
-            crate::parser::Condition::Or(left, right) => {
-                Self::evaluate_condition(ctx, left, pre_resolved)
-                    || Self::evaluate_condition(ctx, right, pre_resolved)
-            }
-            crate::parser::Condition::Eq(left, right) => {
-                let left_str = Self::evaluate_condition_value_string(ctx, left, pre_resolved);
-                let right_str = Self::evaluate_condition_value_string(ctx, right, pre_resolved);
-                left_str == right_str
-            }
-            crate::parser::Condition::NotEq(left, right) => {
-                let left_str = Self::evaluate_condition_value_string(ctx, left, pre_resolved);
-                let right_str = Self::evaluate_condition_value_string(ctx, right, pre_resolved);
-                left_str != right_str
-            }
-            crate::parser::Condition::Gt(left, right) => {
-                Self::evaluate_numeric(ctx, left, right, pre_resolved, |l, r| l > r)
-            }
-            crate::parser::Condition::Lt(left, right) => {
-                Self::evaluate_numeric(ctx, left, right, pre_resolved, |l, r| l < r)
-            }
-            crate::parser::Condition::GtEq(left, right) => {
-                Self::evaluate_numeric(ctx, left, right, pre_resolved, |l, r| l >= r)
-            }
-            crate::parser::Condition::LtEq(left, right) => {
-                Self::evaluate_numeric(ctx, left, right, pre_resolved, |l, r| l <= r)
-            }
-        }
-    }
-
-    #[inline]
-    #[allow(clippy::too_many_arguments)]
-    fn check_will_vacate(
-        ctx: &RenderContext,
-        effective_viewer: &str,
-        recent_borrow: &[crate::models::RecentEntity],
-        future_keys: &[&str],
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-        other_key: &str,
-        other_entity: &dyn TemplateEntity,
-        name: &str,
-        other_name: &str,
-    ) -> bool {
-        if let Some(other_long) = other_entity.long_display_name_for(effective_viewer)
-            && other_long != name
-        {
-            let mut other_short_collisions = 0;
-            let mut other_long_collisions = 0;
-
-            let mut check_entity = |eval_key: &str, eval_entity: &dyn TemplateEntity| {
-                if eval_key != other_key {
-                    let eval_short_name = eval_entity.display_name_for(effective_viewer);
-                    if eval_short_name == other_name {
-                        other_short_collisions += 1;
-                    }
-
-                    let mut is_long_col = eval_short_name == other_long;
-                    if !is_long_col
-                        && eval_short_name == name
-                        && eval_entity
-                            .long_display_name_for(effective_viewer)
-                            .as_deref()
-                            == Some(other_long.as_ref())
-                    {
-                        is_long_col = true;
-                    }
-                    if is_long_col {
-                        other_long_collisions += 1;
-                    }
-                }
-            };
-
-            for r in recent_borrow {
-                if let Ok(eval_entity) = Self::get_entity(ctx, &r.key) {
-                    check_entity(&r.key, eval_entity);
-                }
-            }
-            for &fk in future_keys {
-                if !recent_borrow.iter().any(|r| r.key == fk)
-                    && let Some(&eval_entity) = pre_resolved.get(fk)
-                {
-                    check_entity(fk, eval_entity);
-                }
-            }
-
-            if other_long_collisions < other_short_collisions {
-                return true;
-            }
-        }
-        false
-    }
-
-    fn try_adjective_disambiguation<'a>(
-        entity: &dyn TemplateEntity,
-        colliders: &[&dyn TemplateEntity],
-        base_name: &str,
-        limit: usize,
-    ) -> Option<(std::borrow::Cow<'a, str>, bool)> {
-        let entity_adjs = entity.adjectives().filter(|adjs| !adjs.is_empty())?;
-
-        // Prevent bitshift overflow by capping the limit to the bit-width of our counter
-        let safe_limit = limit.min((u64::BITS - 1) as usize);
-        if safe_limit == 0 {
-            return None;
-        }
-
-        // Filter out adjectives that are shared by ALL colliders, as they provide zero disambiguation value.
-        let mut searchable_adjs = Vec::with_capacity(entity_adjs.len().min(safe_limit));
-        for &adj in entity_adjs {
-            let shared_by_all = colliders.iter().all(|other| {
-                other
-                    .adjectives()
-                    .is_some_and(|other_adjs| other_adjs.contains(&adj))
-            });
-
-            if !shared_by_all {
-                searchable_adjs.push(adj);
-                if searchable_adjs.len() == safe_limit {
-                    break;
-                }
-            }
-        }
-
-        if searchable_adjs.is_empty() {
-            return None;
-        }
-
-        let mut best_combo: Option<Vec<&str>> = None;
-        let mut min_colliders = colliders.len();
-
-        let mut subset = Vec::with_capacity(searchable_adjs.len());
-
-        // Iterate through all non-empty subsets of the entity's adjectives to find the smallest set
-        // that minimizes the number of colliders sharing those adjectives.
-        for i in 1_u64..(1_u64 << searchable_adjs.len()) {
-            subset.clear();
-            for (j, &adj) in searchable_adjs.iter().enumerate() {
-                if (i >> j) & 1 == 1 {
-                    subset.push(adj);
-                }
-            }
-
-            let current_colliders = colliders
-                .iter()
-                .filter(|other| {
-                    if let Some(other_adjs) = other.adjectives() {
-                        subset.iter().all(|&adj| other_adjs.contains(&adj))
-                    } else {
-                        false
-                    }
-                })
-                .count();
-
-            if current_colliders < min_colliders {
-                min_colliders = current_colliders;
-                best_combo = Some(subset.clone());
-            } else if current_colliders == min_colliders {
-                // Prefer the smaller subset if it yields the same number of colliders.
-                if let Some(best) = &best_combo
-                    && subset.len() < best.len()
-                {
-                    best_combo = Some(subset.clone());
-                }
-            }
-        }
-
-        if let Some(combo) = best_combo {
-            // To maintain a stable order, we sort the adjectives before joining.
-            let mut sorted_combo = combo;
-            sorted_combo.sort_unstable();
-            let prefix = sorted_combo.join(" ");
-            Some((
-                std::borrow::Cow::Owned(format!("{prefix} {base_name}")),
-                min_colliders > 0,
-            ))
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    #[allow(clippy::too_many_arguments)]
-    fn count_long_name_collisions(
-        ctx: &RenderContext,
-        effective_viewer: &str,
-        recent_borrow: &[crate::models::RecentEntity],
-        future_keys: &[&str],
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-        params_key: &str,
-        short_name: &str,
-        long_name: &str,
-    ) -> usize {
-        let mut long_collisions = 0;
-
-        // Verify how many times the long name collides
-        let mut check_long = |other_key: &str, other_entity: &dyn TemplateEntity| {
-            if other_key != params_key {
-                let other_short = other_entity.display_name_for(effective_viewer);
-                // Only consider the other entity's long name if its short name is in the
-                // exact same collision group as our entity's short name (preventing phantoms).
-                if other_short == long_name
-                    || (other_short == short_name
-                        && other_entity
-                            .long_display_name_for(effective_viewer)
-                            .as_deref()
-                            == Some(long_name))
-                {
-                    long_collisions += 1;
-                }
-            }
-        };
-
-        for r in recent_borrow {
-            if let Ok(other_entity) = Self::get_entity(ctx, &r.key) {
-                check_long(&r.key, other_entity);
-            }
-        }
-        for &fk in future_keys {
-            if !recent_borrow.iter().any(|r| r.key == fk)
-                && let Some(&other_entity) = pre_resolved.get(fk)
-            {
-                check_long(fk, other_entity);
-            }
-        }
-
-        long_collisions
-    }
-
-    #[inline]
-    fn resolve_display_name<'a>(
-        ctx: &'a RenderContext,
-        entity: &'a dyn TemplateEntity,
-        params: &EntityRefParams<'_>,
-        effective_viewer: &str,
-        future_keys: &[&str],
-        pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
-    ) -> (std::borrow::Cow<'a, str>, Option<usize>) {
-        let mut name = entity.display_name_for(effective_viewer);
-        let mut name_collision = false;
-
-        if !params.flags.no_smart() {
-            let mut short_collisions = 0;
-            let mut unresolved_colliders: Vec<&'a dyn TemplateEntity> = Vec::new();
-            let recent_borrow = ctx.recent_entities.borrow();
-
-            // We use a closure here to iterate over the live `recent_borrow` and `future_keys`.
-            // Evaluating collisions dynamically avoids allocations while ensuring we accurately
-            // catch entities that were just introduced in this template, maintaining strict
-            // left-to-right chronological accuracy.
-            let mut check_collision = |other_key: &str, other_entity: &'a dyn TemplateEntity| {
-                if other_key != params.key
-                    && other_entity.display_name_for(effective_viewer) == name
-                {
-                    short_collisions += 1; // This counts all collisions, even those that will be resolved by long names.
-
-                    // Determine if this other entity will vacate the short name by using its own long name
-                    if !Self::check_will_vacate(
-                        ctx,
-                        effective_viewer,
-                        &recent_borrow,
-                        future_keys,
-                        pre_resolved,
-                        other_key,
-                        other_entity,
-                        name.as_ref(),
-                        // We pass `name` again for `other_name` because the `if` condition above
-                        // proved they are identical, saving us from binding a temporary variable.
-                        name.as_ref(),
-                    ) {
-                        unresolved_colliders.push(other_entity);
-                    }
-                }
-            };
-
-            for r in recent_borrow.iter() {
-                if let Ok(other_entity) = Self::get_entity(ctx, &r.key) {
-                    check_collision(&r.key, other_entity);
-                }
-            }
-            for &fk in future_keys {
-                if !recent_borrow.iter().any(|r| r.key == fk)
-                    && let Some(&other_entity) = pre_resolved.get(fk)
-                {
-                    check_collision(fk, other_entity);
-                }
-            }
-
-            name_collision = !unresolved_colliders.is_empty();
-
-            if short_collisions > 0
-                && let Some(long_name) = entity.long_display_name_for(effective_viewer)
-                && long_name != name
-            {
-                let long_collisions = Self::count_long_name_collisions(
-                    ctx,
-                    effective_viewer,
-                    &recent_borrow,
-                    future_keys,
-                    pre_resolved,
-                    params.key,
-                    name.as_ref(),
-                    long_name.as_ref(),
-                );
-
-                // If the long name is strictly more specific (has fewer collisions), use it!
-                if long_collisions < short_collisions {
-                    name = long_name;
-                    name_collision = long_collisions > 0;
-                }
-            }
-
-            // If a collision still exists, try to disambiguate by prepending unique adjectives.
-            if name_collision
-                && !entity.is_proper_noun_for(effective_viewer)
-                && let Some((adj_name, still_collides)) = Self::try_adjective_disambiguation(
-                    entity,
-                    &unresolved_colliders,
-                    &name,
-                    ctx.adjective_disambiguation_limit,
-                )
-            {
-                name = adj_name;
-                name_collision = still_collides;
-            }
-        }
-
-        let mut ordinal = None;
-
-        if !params.flags.no_smart() {
-            let namespace = params
-                .owner_key
-                .or_else(|| params.key.rsplit_once(TAG_PROPERTY_SEP).map(|(p, _)| p));
-            let ordinal_group_name = if let Some(ns) = namespace {
-                std::borrow::Cow::Owned(format!("{ns}::{name}"))
-            } else {
-                std::borrow::Cow::Borrowed(name.as_ref())
-            };
-
-            ordinal =
-                Self::assign_ordinal(ctx, ordinal_group_name.as_ref(), params.key, name_collision);
-        }
-
-        (name, ordinal)
-    }
-
-    #[inline]
-    fn assign_ordinal(
-        ctx: &RenderContext,
-        name: &str,
-        key: &str,
-        name_collision: bool,
-    ) -> Option<usize> {
-        let mut ordinals = ctx.ordinals.borrow_mut();
-        ctx.clear_target_cache();
-        if !ordinals.contains_key(name) {
-            ordinals.insert(
-                name.to_string(),
-                crate::models::OrdinalState {
-                    next_ordinal: 1,
-                    members: HashMap::new(),
-                },
-            );
-        }
-
-        if let Some(state) = ordinals.get_mut(name) {
-            if name_collision {
-                if !state.members.contains_key(key) {
-                    let o = state.next_ordinal;
-                    state.next_ordinal += 1;
-                    state.members.insert(key.to_string(), o);
-                }
-                if let Some(&ord) = state.members.get(key) {
-                    return Some(ord);
-                }
-            } else {
-                state.members.clear();
-                state.members.insert(key.to_string(), 1);
-                state.next_ordinal = 2;
-            }
-        }
-        None
-    }
-
     fn render_entity_ref<'a>(
         ctx: &'a RenderContext,
         raw_output: &mut String,
@@ -849,7 +280,7 @@ impl PerspectiveEngine {
         let entity = pre_resolved
             .get(params.key)
             .copied()
-            .map_or_else(|| Self::get_entity(ctx, params.key), Ok)?;
+            .map_or_else(|| crate::evaluator::get_entity(ctx, params.key), Ok)?;
 
         let effective_viewer = effective_viewer_id(ctx, params.flags.force_3rd_person());
 
@@ -987,7 +418,7 @@ impl PerspectiveEngine {
         let owner_entity = pre_resolved
             .get(owner_key)
             .copied()
-            .map_or_else(|| Self::get_entity(ctx, owner_key), Ok)?;
+            .map_or_else(|| crate::evaluator::get_entity(ctx, owner_key), Ok)?;
 
         let owner_viewer = effective_viewer_id(ctx, params.owner_flags.force_3rd_person());
 
@@ -1089,29 +520,6 @@ impl PerspectiveEngine {
     }
 
     #[inline]
-    fn determine_group_singular_gender(members: &[&dyn TemplateEntity]) -> crate::models::Gender {
-        let mut flat = Vec::new();
-        crate::models::flatten_group(members, &mut flat, 0);
-        let mut shared = None;
-        for m in flat {
-            let g = m.gender();
-            let singular_g = if g == crate::models::Gender::Plural {
-                crate::models::Gender::Neutral
-            } else {
-                g
-            };
-            if let Some(s) = shared {
-                if s != singular_g {
-                    return crate::models::Gender::Neutral;
-                }
-            } else {
-                shared = Some(singular_g);
-            }
-        }
-        shared.unwrap_or(crate::models::Gender::Neutral)
-    }
-
-    #[inline]
     fn try_render_pronoun<'a>(
         ctx: &RenderContext,
         raw_output: &mut String,
@@ -1140,7 +548,7 @@ impl PerspectiveEngine {
         if params.flags.force_singular() || extract_member {
             is_plural = false;
             if let Some(members) = entity.group_members() {
-                effective_gender = Self::determine_group_singular_gender(members);
+                effective_gender = groups::determine_group_singular_gender(members);
             } else if effective_gender == crate::models::Gender::Plural {
                 effective_gender = crate::models::Gender::Neutral;
             }
@@ -1148,7 +556,7 @@ impl PerspectiveEngine {
 
         let is_viewer = entity.contains_viewer(effective_viewer) && !extract_member;
         let is_active_subject =
-            Self::check_is_active_subject(ctx, entity, params.key, pre_resolved);
+            anaphora::check_is_active_subject(ctx, entity, params.key, pre_resolved);
 
         if is_active_subject && p_type == "obj" {
             actual_p_type = Some("reflex");
@@ -1179,8 +587,8 @@ impl PerspectiveEngine {
 
         if can_use_pronoun {
             if !already_seen {
-                update_memory(&ctx.last_mentioned, params.key);
-                track_recent_entity(ctx, params.key, entity, params.adjectives, None);
+                anaphora::update_memory(&ctx.last_mentioned, params.key);
+                anaphora::track_recent_entity(ctx, params.key, entity, params.adjectives, None);
             }
             let pronoun = resolve_pronoun(
                 effective_gender,
@@ -1260,7 +668,13 @@ impl PerspectiveEngine {
         if !can_use_pronoun
             && already_seen
             && !active_flags.prefer_noun()
-            && !Self::is_pronoun_ambiguous(ctx, key, effective_gender, pronoun_ctx, active_flags)
+            && !anaphora::is_pronoun_ambiguous(
+                ctx,
+                key,
+                effective_gender,
+                pronoun_ctx,
+                active_flags,
+            )
         {
             can_use_pronoun = true;
         }
@@ -1283,7 +697,7 @@ impl PerspectiveEngine {
         future_keys: &[&str],
         pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
     ) {
-        let (name, ordinal) = Self::resolve_display_name(
+        let (name, ordinal) = disambiguation::resolve_display_name(
             ctx,
             entity,
             params,
@@ -1317,8 +731,8 @@ impl PerspectiveEngine {
             }
         }
 
-        update_memory(&ctx.last_mentioned, params.key);
-        track_recent_entity(
+        anaphora::update_memory(&ctx.last_mentioned, params.key);
+        anaphora::track_recent_entity(
             ctx,
             params.key,
             entity,
@@ -1344,7 +758,7 @@ impl PerspectiveEngine {
 
         // --- Handle Groups / Distributed Lists ---
         if let Some(members) = entity.group_members() {
-            Self::render_group_entity(
+            groups::render_group_entity(
                 ctx,
                 raw_output,
                 entity,
@@ -1472,345 +886,7 @@ impl PerspectiveEngine {
     }
 
     #[inline]
-    #[allow(clippy::too_many_arguments)]
-    fn render_group_entity<'a>(
-        ctx: &'a RenderContext,
-        raw_output: &mut String,
-        entity: &'a dyn TemplateEntity,
-        members: &[&'a dyn TemplateEntity],
-        effective_viewer: &str,
-        params: &EntityRefParams<'_>,
-        cap_whole: bool,
-        after_possessive: bool,
-        pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
-    ) {
-        let (viewer_entity, visible) = crate::models::partition_group(members, effective_viewer);
-
-        let total_visible = visible.len() + usize::from(viewer_entity.is_some());
-        if total_visible == 0 {
-            return;
-        }
-
-        let active_subject_entity = ctx
-            .active_subject
-            .borrow()
-            .as_deref()
-            .and_then(|active_key| {
-                pre_resolved
-                    .get(active_key)
-                    .copied()
-                    .map_or_else(|| Self::get_entity(ctx, active_key).ok(), Some)
-            });
-
-        let viewer_is_active_subject =
-            active_subject_entity.is_some_and(|active| active.contains_viewer(effective_viewer));
-        let is_objective = params.p_type == Some("obj");
-
-        let mut ends_with_possessive_pronoun = false;
-        let mut decomposed_we = false;
-        let mut formatted_names = Vec::with_capacity(total_visible + 1);
-
-        if let Some(viewer) = viewer_entity
-            && let Some(prefix) = Self::format_group_viewer_prefix(
-                ctx,
-                viewer,
-                params,
-                visible.is_empty(),
-                viewer_is_active_subject,
-                is_objective,
-                &mut ends_with_possessive_pronoun,
-                &mut decomposed_we,
-            )
-        {
-            formatted_names.push(prefix);
-        }
-
-        let will_append_my = viewer_entity.is_some_and(|viewer| {
-            ctx.stance == crate::models::ActorStance::FirstPerson
-                && (!viewer.is_plural() || decomposed_we)
-        });
-
-        let distribute_possessives = viewer_entity.is_some() && params.flags.is_possessive();
-
-        let lower_article_storage = params.article.map(str::to_lowercase);
-        let mut first_visible_item = viewer_entity.is_none();
-        for (member, name) in visible {
-            let article_to_use = if first_visible_item {
-                params.article
-            } else {
-                lower_article_storage.as_deref()
-            };
-
-            let member_is_active_subj =
-                active_subject_entity.is_some_and(|active| is_same_entity(active, member));
-
-            let mut flags = GroupMemberFlags::empty();
-            flags.set(GroupMemberFlags::AFTER_POSSESSIVE, after_possessive);
-            flags.set(GroupMemberFlags::FIRST_VISIBLE_ITEM, first_visible_item);
-            flags.set(
-                GroupMemberFlags::DISTRIBUTE_POSSESSIVES,
-                distribute_possessives,
-            );
-            flags.set(
-                GroupMemberFlags::IS_REFLEXIVE,
-                is_objective && member_is_active_subj,
-            );
-
-            let config = GroupMemberFormatConfig {
-                flags,
-                article_to_use,
-            };
-
-            formatted_names.push(Self::format_group_member(
-                ctx,
-                entity,
-                member,
-                name,
-                effective_viewer,
-                params,
-                &config,
-            ));
-            first_visible_item = false;
-        }
-
-        if will_append_my {
-            let suffix = Self::format_group_viewer_suffix(
-                params.flags.is_possessive(),
-                viewer_is_active_subject,
-                is_objective,
-                &mut ends_with_possessive_pronoun,
-            );
-            formatted_names.push(suffix);
-        }
-
-        let conjunction = if params.flags.extract_group_member() {
-            "or"
-        } else {
-            "and"
-        };
-        let list_str = crate::grammar::format_oxford_list(formatted_names, conjunction);
-
-        let mut final_str = list_str.into_owned();
-        if params.flags.is_possessive() && !ends_with_possessive_pronoun && !distribute_possessives
-        {
-            final_str.push_str(Self::get_possessive_suffix(&final_str, entity.is_plural()));
-        }
-
-        push_capitalized_if(raw_output, &final_str, cap_whole);
-    }
-
-    #[inline]
-    #[allow(clippy::too_many_arguments)]
-    fn format_group_viewer_prefix(
-        ctx: &RenderContext,
-        viewer: &dyn TemplateEntity,
-        params: &EntityRefParams<'_>,
-        visible_is_empty: bool,
-        viewer_is_active_subject: bool,
-        is_objective: bool,
-        ends_with_possessive_pronoun: &mut bool,
-        decomposed_we: &mut bool,
-    ) -> Option<std::borrow::Cow<'static, str>> {
-        if ctx.stance == crate::models::ActorStance::SecondPerson {
-            if params.flags.is_possessive() {
-                if visible_is_empty {
-                    *ends_with_possessive_pronoun = true;
-                }
-                return Some(std::borrow::Cow::Borrowed("your"));
-            } else if is_objective && viewer_is_active_subject {
-                let reflex = resolve_pronoun(
-                    crate::models::Gender::Neutral,
-                    "reflex",
-                    true,
-                    viewer.is_plural(),
-                    ctx.stance,
-                )
-                .unwrap_or("yourself");
-                return Some(std::borrow::Cow::Borrowed(reflex));
-            }
-            return Some(std::borrow::Cow::Borrowed("you"));
-        } else if ctx.stance == crate::models::ActorStance::FirstPerson && viewer.is_plural() {
-            if visible_is_empty {
-                if params.flags.is_possessive() {
-                    *ends_with_possessive_pronoun = true;
-                    return Some(std::borrow::Cow::Borrowed("our"));
-                } else if is_objective {
-                    if viewer_is_active_subject {
-                        let reflex = resolve_pronoun(
-                            crate::models::Gender::Neutral,
-                            "reflex",
-                            true,
-                            true,
-                            ctx.stance,
-                        )
-                        .unwrap_or("ourselves");
-                        return Some(std::borrow::Cow::Borrowed(reflex));
-                    }
-                    return Some(std::borrow::Cow::Borrowed("us"));
-                }
-                return Some(std::borrow::Cow::Borrowed("we"));
-            }
-
-            *decomposed_we = true;
-            if params.flags.is_possessive() {
-                return Some(std::borrow::Cow::Borrowed("your"));
-            } else if is_objective && viewer_is_active_subject {
-                let reflex = resolve_pronoun(
-                    crate::models::Gender::Neutral,
-                    "reflex",
-                    true,
-                    false,
-                    crate::models::ActorStance::SecondPerson,
-                )
-                .unwrap_or("yourself");
-                return Some(std::borrow::Cow::Borrowed(reflex));
-            }
-            return Some(std::borrow::Cow::Borrowed("you"));
-        }
-        None
-    }
-
-    #[inline]
-    fn format_group_viewer_suffix(
-        is_possessive: bool,
-        viewer_is_active_subject: bool,
-        is_objective: bool,
-        ends_with_possessive_pronoun: &mut bool,
-    ) -> std::borrow::Cow<'static, str> {
-        if is_possessive {
-            *ends_with_possessive_pronoun = true;
-            std::borrow::Cow::Borrowed("my")
-        } else if is_objective {
-            if viewer_is_active_subject {
-                let reflex = resolve_pronoun(
-                    crate::models::Gender::Neutral,
-                    "reflex",
-                    true,
-                    false,
-                    crate::models::ActorStance::FirstPerson,
-                )
-                .unwrap_or("myself");
-                std::borrow::Cow::Borrowed(reflex)
-            } else {
-                std::borrow::Cow::Borrowed("me")
-            }
-        } else {
-            std::borrow::Cow::Borrowed("I")
-        }
-    }
-
-    #[inline]
-    #[allow(clippy::too_many_arguments)]
-    fn format_group_member<'a>(
-        ctx: &RenderContext,
-        entity: &dyn TemplateEntity,
-        member: &'a dyn TemplateEntity,
-        name: std::borrow::Cow<'a, str>,
-        effective_viewer: &str,
-        params: &EntityRefParams<'_>,
-        config: &GroupMemberFormatConfig<'_>,
-    ) -> std::borrow::Cow<'a, str> {
-        let first_visible_item = config.flags.contains(GroupMemberFlags::FIRST_VISIBLE_ITEM);
-
-        if config.flags.contains(GroupMemberFlags::IS_REFLEXIVE) {
-            let reflex = resolve_pronoun(
-                member.gender(),
-                "reflex",
-                false,
-                member.is_plural(),
-                ctx.stance,
-            )
-            .unwrap_or("itself");
-            let mut final_name = if (params.flags.is_capitalized()
-                || params.flags.article_capitalized())
-                && first_visible_item
-            {
-                std::borrow::Cow::Owned(crate::grammar::capitalize_first(reflex))
-            } else {
-                std::borrow::Cow::Borrowed(reflex)
-            };
-
-            if config
-                .flags
-                .contains(GroupMemberFlags::DISTRIBUTE_POSSESSIVES)
-            {
-                let suffix = Self::get_possessive_suffix(&final_name, member.is_plural());
-                let mut owned = final_name.into_owned();
-                owned.push_str(suffix);
-                final_name = std::borrow::Cow::Owned(owned);
-            }
-            return final_name;
-        }
-
-        let mut article_flags = crate::grammar::ArticleFlags::empty();
-        article_flags.set(
-            crate::grammar::ArticleFlags::IS_PROPER_NOUN,
-            member.is_proper_noun_for(effective_viewer),
-        );
-        article_flags.set(crate::grammar::ArticleFlags::IS_PLURAL, member.is_plural());
-        article_flags.set(
-            crate::grammar::ArticleFlags::FORCE_ARTICLE,
-            params.flags.force_article(),
-        );
-        article_flags.set(
-            crate::grammar::ArticleFlags::AFTER_POSSESSIVE,
-            config.flags.contains(GroupMemberFlags::AFTER_POSSESSIVE),
-        );
-        article_flags.set(
-            crate::grammar::ArticleFlags::IS_CAPITALIZED,
-            params.flags.article_capitalized() && first_visible_item,
-        );
-
-        let mut adj_prefix = String::new();
-        if first_visible_item
-            && let Some(adj) = params.adjectives
-            && !adj.is_empty()
-        {
-            if params.flags.contains(TagFlags::ALL_CAPS) {
-                crate::typography::apply_all_caps(adj, &mut adj_prefix);
-            } else {
-                adj_prefix.push_str(adj);
-            }
-            adj_prefix.push(' ');
-        }
-
-        let mut final_name = if let Some(resolved_art) = config.article_to_use.and_then(|art| {
-            resolve_article(
-                art,
-                &name,
-                params.ordinal,
-                entity.collective_noun(),
-                ctx.ordinal_word_threshold,
-                article_flags,
-            )
-        }) {
-            std::borrow::Cow::Owned(format!("{}{adj_prefix}{name}", resolved_art.as_ref()))
-        } else if !adj_prefix.is_empty() {
-            let cap_adj = if params.flags.article_capitalized() && first_visible_item {
-                crate::grammar::capitalize_first(&adj_prefix)
-            } else {
-                adj_prefix
-            };
-            std::borrow::Cow::Owned(format!("{cap_adj}{name}"))
-        } else {
-            name
-        };
-
-        if config
-            .flags
-            .contains(GroupMemberFlags::DISTRIBUTE_POSSESSIVES)
-        {
-            let suffix = Self::get_possessive_suffix(&final_name, member.is_plural());
-            let mut owned = final_name.into_owned();
-            owned.push_str(suffix);
-            final_name = std::borrow::Cow::Owned(owned);
-        }
-
-        final_name
-    }
-
-    #[inline]
-    fn get_last_visible_char(input: &str) -> Option<char> {
+    pub(crate) fn get_last_visible_char(input: &str) -> Option<char> {
         #[cfg(not(any(feature = "mxp", feature = "msp", feature = "ansi")))]
         {
             input.trim_end().chars().next_back()
@@ -1844,74 +920,11 @@ impl PerspectiveEngine {
     }
 
     #[inline]
-    fn get_possessive_suffix(name: &str, is_plural: bool) -> &'static str {
+    pub(crate) fn get_possessive_suffix(name: &str, is_plural: bool) -> &'static str {
         if matches!(Self::get_last_visible_char(name), Some('s' | 'S')) && is_plural {
             return "'";
         }
         MOD_POSSESSIVE
-    }
-
-    #[inline]
-    fn check_is_active_subject(
-        ctx: &RenderContext,
-        entity: &dyn TemplateEntity,
-        key: &str,
-        pre_resolved: &HashMap<&str, &dyn TemplateEntity>,
-    ) -> bool {
-        if ctx.active_subject.borrow().as_deref() == Some(key) {
-            return true;
-        }
-        if let Some(active_key) = ctx.active_subject.borrow().as_deref()
-            && let Ok(active_entity) = pre_resolved
-                .get(active_key)
-                .copied()
-                .map_or_else(|| Self::get_entity(ctx, active_key), Ok)
-        {
-            // Prevent false positives when a sub-property is located at memory offset 0
-            // of its parent struct, which causes their data pointers to be identical.
-            if is_sub_property_path(active_key, key) || is_sub_property_path(key, active_key) {
-                return false;
-            }
-            return is_same_entity(entity, active_entity);
-        }
-        false
-    }
-
-    #[inline]
-    fn is_pronoun_ambiguous(
-        ctx: &RenderContext,
-        key: &str,
-        effective_gender: crate::models::Gender,
-        pronoun_ctx: PronounContextFlags,
-        flags: TagFlags,
-    ) -> bool {
-        let is_plural = pronoun_ctx.contains(PronounContextFlags::IS_PLURAL);
-        for other in ctx.recent_entities.borrow().iter() {
-            if other.key != key {
-                let other_is_viewer = if flags.force_3rd_person()
-                    || ctx.stance == crate::models::ActorStance::ThirdPerson
-                {
-                    other
-                        .flags
-                        .contains(crate::models::RecentEntityFlags::IS_VIEWER_FORCED)
-                } else {
-                    other
-                        .flags
-                        .contains(crate::models::RecentEntityFlags::IS_VIEWER_NORMAL)
-                };
-
-                if !other_is_viewer
-                    && effective_gender == other.gender
-                    && is_plural
-                        == other
-                            .flags
-                            .contains(crate::models::RecentEntityFlags::IS_PLURAL)
-                {
-                    return true;
-                }
-            }
-        }
-        false
     }
 
     fn render_variable_ref<'a>(
@@ -1922,28 +935,13 @@ impl PerspectiveEngine {
         flags: TagFlags,
         pre_resolved: &HashMap<&str, &'a dyn TemplateEntity>,
     ) -> Result<(), String> {
-        if let Some((ent, prop)) = key.rsplit_once('.') {
-            let root_key = ent.split_once('.').map_or(ent, |(r, _)| r);
-            if ctx.entities.contains_key(root_key) {
-                if let Ok(entity) = pre_resolved
-                    .get(ent)
-                    .copied()
-                    .map_or_else(|| Self::get_entity(ctx, ent), Ok)
-                    && let Some(val) = entity.get_string_property(prop)
-                {
-                    Self::apply_variable_formatting(val.as_ref(), raw_output, flags);
-                    return Ok(());
-                }
-                if let Some(fb) = fallback {
-                    Self::apply_variable_formatting(fb, raw_output, flags);
-                    return Ok(());
-                }
-                tracing::error!("Missing string property '{prop}' on entity '{ent}'");
-                return Err(format!(
-                    "Missing string property '{prop}' on entity '{ent}'"
-                ));
-            }
+        if let Some(val) =
+            crate::evaluator::resolve_entity_property(ctx, key, fallback, pre_resolved)?
+        {
+            Self::apply_variable_formatting(val.as_ref(), raw_output, flags);
+            return Ok(());
         }
+
         if let Some(values) = ctx.variables.get(key) {
             if values.is_empty() {
                 return Ok(());
@@ -2013,7 +1011,7 @@ impl PerspectiveEngine {
 
         let resolved_subject_key = subject_key
             .as_ref()
-            .map(|s| Self::resolve_tag_segment(ctx, s, pre_resolved))
+            .map(|s| crate::evaluator::resolve_tag_segment(ctx, s, pre_resolved))
             .transpose()?;
 
         // Explicitly bind the verb to its subject to solve passive voice / compound subjects
@@ -2022,11 +1020,11 @@ impl PerspectiveEngine {
             let entity = pre_resolved
                 .get(key_str)
                 .copied()
-                .map_or_else(|| Self::get_entity(ctx, key_str), Ok)?;
+                .map_or_else(|| crate::evaluator::get_entity(ctx, key_str), Ok)?;
             let effective_viewer = effective_viewer_id(ctx, flags.force_3rd_person());
-            update_memory(&ctx.active_subject, key_str);
-            update_memory(&ctx.last_mentioned, key_str);
-            track_recent_entity(ctx, key_str, entity, None, None);
+            anaphora::update_memory(&ctx.active_subject, key_str);
+            anaphora::update_memory(&ctx.last_mentioned, key_str);
+            anaphora::track_recent_entity(ctx, key_str, entity, None, None);
             (
                 entity.contains_viewer(effective_viewer),
                 entity.is_plural(),
@@ -2117,34 +1115,20 @@ impl PerspectiveEngine {
     ) -> Result<(), String> {
         let mut conjugated_verbs = Vec::new();
 
-        if let Some((ent, prop)) = d_key.rsplit_once('.') {
-            let root_key = ent.split_once('.').map_or(ent, |(r, _)| r);
-            if ctx.entities.contains_key(root_key) {
-                let entity = pre_resolved
-                    .get(ent)
-                    .copied()
-                    .map_or_else(|| Self::get_entity(ctx, ent), Ok)?;
-                if let Some(val) = entity.get_string_property(prop) {
-                    let v_lower = val.to_lowercase();
-                    let conjugated = conjugate_verb(
-                        &val,
-                        &v_lower,
-                        flags.is_capitalized(),
-                        is_viewer,
-                        is_plural,
-                        ctx.stance,
-                        ctx.tense,
-                    );
-                    conjugated_verbs.push(conjugated.into_owned().into());
-                } else if let Some(fb) = dynamic_fallback {
-                    conjugated_verbs.push(std::borrow::Cow::Owned(fb.to_string()));
-                } else {
-                    tracing::error!("Missing string property '{prop}' on entity '{ent}'");
-                    return Err(format!(
-                        "Missing string property '{prop}' on entity '{ent}'"
-                    ));
-                }
-            }
+        if let Some(val) =
+            crate::evaluator::resolve_entity_property(ctx, d_key, dynamic_fallback, pre_resolved)?
+        {
+            let v_lower = val.to_lowercase();
+            let conjugated = conjugate_verb(
+                &val,
+                &v_lower,
+                flags.is_capitalized(),
+                is_viewer,
+                is_plural,
+                ctx.stance,
+                ctx.tense,
+            );
+            conjugated_verbs.push(conjugated.into_owned().into());
         }
 
         if conjugated_verbs.is_empty() {
@@ -2189,98 +1173,6 @@ impl PerspectiveEngine {
 }
 
 #[inline]
-fn update_memory(memory: &std::cell::RefCell<Option<String>>, key: &str) {
-    if memory.borrow().as_deref() != Some(key) {
-        *memory.borrow_mut() = Some(key.to_string());
-    }
-}
-
-#[inline]
-fn track_recent_entity(
-    ctx: &RenderContext<'_>,
-    key: &str,
-    entity: &dyn TemplateEntity,
-    adjectives: Option<&str>,
-    resolved_name: Option<&str>,
-) {
-    let mut recents = ctx.recent_entities.borrow_mut();
-    ctx.clear_target_cache();
-
-    let mut new_adjs = Vec::new();
-    if let Some(adj_str) = adjectives {
-        let stripped = crate::typography::strip_all_protocol_tags(adj_str);
-        for word in stripped.split_whitespace() {
-            let clean = word.trim().to_lowercase();
-            if !clean.is_empty() {
-                new_adjs.push(clean);
-            }
-        }
-    }
-
-    // Move to the back to represent the most recently used (LRU)
-    if let Some(pos) = recents.iter().position(|r| r.key == key) {
-        let mut item = recents.remove(pos);
-
-        // Refresh grammatical properties in case the underlying entity mutated (e.g., GroupEntities)
-        item.gender = entity.gender();
-        item.flags.set(
-            crate::models::RecentEntityFlags::IS_PLURAL,
-            entity.is_plural(),
-        );
-        item.flags.set(
-            crate::models::RecentEntityFlags::IS_VIEWER_NORMAL,
-            entity.contains_viewer(ctx.viewer_id),
-        );
-        item.flags.set(
-            crate::models::RecentEntityFlags::IS_VIEWER_FORCED,
-            entity.contains_viewer(crate::models::NULL_VIEWER),
-        );
-
-        for adj in new_adjs {
-            if !item.adjectives.contains(&adj) {
-                item.adjectives.push(adj);
-            }
-        }
-        if let Some(rn) = resolved_name {
-            item.resolved_name = Some(rn.to_string());
-        }
-        recents.push(item);
-    } else {
-        let mut flags = crate::models::RecentEntityFlags::empty();
-        flags.set(
-            crate::models::RecentEntityFlags::IS_PLURAL,
-            entity.is_plural(),
-        );
-        flags.set(
-            crate::models::RecentEntityFlags::IS_VIEWER_NORMAL,
-            entity.contains_viewer(ctx.viewer_id),
-        );
-        flags.set(
-            crate::models::RecentEntityFlags::IS_VIEWER_FORCED,
-            entity.contains_viewer(crate::models::NULL_VIEWER),
-        );
-
-        recents.push(crate::models::RecentEntity {
-            key: key.to_string(),
-            gender: entity.gender(),
-            flags,
-            adjectives: new_adjs,
-            resolved_name: resolved_name.map(ToString::to_string),
-        });
-    }
-
-    // Enforce the anaphora memory capacity limit
-    let mut last_mentioned = ctx.last_mentioned.borrow_mut();
-    let mut active_subject = ctx.active_subject.borrow_mut();
-    crate::models::enforce_anaphora_limit(
-        ctx.anaphora_limit,
-        &mut recents,
-        &mut last_mentioned,
-        &mut active_subject,
-    );
-}
-
-#[inline]
 fn should_capitalize_whole_tag(params: &EntityRefParams<'_>) -> bool {
     params.flags.is_capitalized() || params.flags.article_capitalized()
 }
@@ -2316,13 +1208,6 @@ fn viewer_name(stance: crate::models::ActorStance, flags: ViewerNameFlags) -> &'
         },
         crate::models::ActorStance::ThirdPerson => "",
     }
-}
-
-#[inline]
-fn is_sub_property_path(parent: &str, child: &str) -> bool {
-    child.starts_with(parent)
-        && child.len() > parent.len()
-        && child.as_bytes().get(parent.len()) == Some(&b'.')
 }
 
 #[inline]
