+++ src/parser.rs
@@ -1,5 +1,12 @@
 use std::collections::HashSet;
 
+/// Parsing logic for boolean expressions and conditionals.
+pub mod expressions;
+pub use expressions::{Condition, ConditionValue};
+/// Utility functions for text processing and tag validation.
+pub mod utils;
+pub(crate) use utils::*;
+
 #[cfg(any(feature = "mxp", feature = "msp", feature = "ansi"))]
 use crate::typography::{has_protocol_tags, skip_protocol_tags};
 
@@ -48,379 +55,14 @@ impl TagSegment {
     pub(crate) fn parse(s: &str) -> Self {
         let s = s.trim();
         if let Some(var) = s.strip_prefix('$') {
-            if let Some((key, fallback)) = var.split_once("??") {
-                Self::Variable {
-                    key: key.trim().to_lowercase(),
-                    fallback: Some(unescape_string(strip_outer_quotes(fallback))),
-                }
-            } else {
-                Self::Variable {
-                    key: var.trim().to_lowercase(),
-                    fallback: None,
-                }
+            let (key, fallback) = extract_variable_fallback(var);
+            Self::Variable {
+                key: key.to_lowercase(),
+                fallback,
             }
         } else {
             Self::Literal(s.to_string())
         }
-    }
-}
-
-/// A value evaluated within a conditional statement.
-#[derive(Debug, Clone, PartialEq)]
-pub enum ConditionValue {
-    /// A static string literal (e.g., `"raining"`).
-    Literal(String),
-    /// A numerical float parsed for inequalities.
-    Number(f64),
-    /// A dynamic context variable (e.g., `$weather`).
-    Variable(String),
-    /// A dynamic entity property (e.g., `source.color`).
-    EntityProperty(String, String),
-}
-
-/// A logical condition that can be evaluated dynamically at render time.
-#[derive(Debug, Clone, PartialEq)]
-pub enum Condition {
-    /// Evaluates a single value for truthiness.
-    Value(ConditionValue),
-    /// Inverts the boolean result of an expression.
-    Not(Box<Condition>),
-    /// Evaluates true if both expressions are true.
-    And(Box<Condition>, Box<Condition>),
-    /// Evaluates true if either expression is true.
-    Or(Box<Condition>, Box<Condition>),
-    /// Evaluates equality between two condition values.
-    Eq(ConditionValue, ConditionValue),
-    /// Evaluates inequality between two condition values.
-    NotEq(ConditionValue, ConditionValue),
-    /// Evaluates if the left value is strictly greater than the right value (numerically).
-    Gt(ConditionValue, ConditionValue),
-    /// Evaluates if the left value is strictly less than the right value (numerically).
-    Lt(ConditionValue, ConditionValue),
-    /// Evaluates if the left value is greater than or equal to the right value (numerically).
-    GtEq(ConditionValue, ConditionValue),
-    /// Evaluates if the left value is less than or equal to the right value (numerically).
-    LtEq(ConditionValue, ConditionValue),
-}
-
-impl Condition {
-    pub(crate) fn parse(s: &str, max_depth: usize) -> Result<Self, String> {
-        let tokens = tokenize_expr(s)?;
-        ExprParser::parse(&tokens, max_depth)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum ExprToken {
-    And,
-    Or,
-    Not,
-    LParen,
-    RParen,
-    Eq,
-    NotEq,
-    Gt,
-    Lt,
-    GtEq,
-    LtEq,
-    Val(ConditionValue),
-}
-
-#[allow(clippy::too_many_lines)]
-fn tokenize_expr(s: &str) -> Result<Vec<ExprToken>, String> {
-    let mut tokens = Vec::new();
-    let mut chars = s.char_indices().peekable();
-
-    while let Some(&(i, c)) = chars.peek() {
-        if c.is_whitespace() {
-            chars.next();
-            continue;
-        }
-        match c {
-            '(' => {
-                tokens.push(ExprToken::LParen);
-                chars.next();
-            }
-            ')' => {
-                tokens.push(ExprToken::RParen);
-                chars.next();
-            }
-            '=' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '=').is_some() {
-                    tokens.push(ExprToken::Eq);
-                } else {
-                    return Err(format!("Expected '==' at index {i}"));
-                }
-            }
-            '!' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '=').is_some() {
-                    tokens.push(ExprToken::NotEq);
-                } else {
-                    tokens.push(ExprToken::Not);
-                }
-            }
-            '>' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '=').is_some() {
-                    tokens.push(ExprToken::GtEq);
-                } else {
-                    tokens.push(ExprToken::Gt);
-                }
-            }
-            '<' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '=').is_some() {
-                    tokens.push(ExprToken::LtEq);
-                } else {
-                    tokens.push(ExprToken::Lt);
-                }
-            }
-            '&' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '&').is_some() {
-                    tokens.push(ExprToken::And);
-                } else {
-                    return Err(format!("Expected '&&' at index {i}"));
-                }
-            }
-            '|' => {
-                chars.next();
-                if chars.next_if(|&(_, n)| n == '|').is_some() {
-                    tokens.push(ExprToken::Or);
-                } else {
-                    return Err(format!("Expected '||' at index {i}"));
-                }
-            }
-            '`' => {
-                chars.next();
-                let mut s = String::new();
-                while let Some(&(_, n)) = chars.peek() {
-                    if n == '`' {
-                        chars.next();
-                        break;
-                    }
-                    s.push(n);
-                    chars.next();
-                }
-                tokens.push(ExprToken::Val(ConditionValue::Literal(s)));
-            }
-            '"' | '\'' => {
-                let quote = c;
-                chars.next();
-                let mut s = String::new();
-                while let Some(&(_, n)) = chars.peek() {
-                    if n == '\\' {
-                        chars.next(); // Consume the '\'
-                        if let Some(&(_, escaped_c)) = chars.peek() {
-                            chars.next(); // Consume the escaped character
-                            match escaped_c {
-                                'n' => s.push('\n'),
-                                'r' => s.push('\r'),
-                                't' => s.push('\t'),
-                                'u' => {
-                                    if let Some(&(_, '{')) = chars.peek() {
-                                        chars.next(); // Consume '{'
-                                        let mut hex = String::new();
-                                        while let Some(&(_, hc)) = chars.peek() {
-                                            chars.next();
-                                            if hc == '}' {
-                                                break;
-                                            }
-                                            hex.push(hc);
-                                        }
-                                        if let Ok(code) = u32::from_str_radix(&hex, 16) {
-                                            s.push(
-                                                char::from_u32(code)
-                                                    .unwrap_or(char::REPLACEMENT_CHARACTER),
-                                            );
-                                        } else {
-                                            s.push(char::REPLACEMENT_CHARACTER);
-                                        }
-                                    } else {
-                                        s.push('u');
-                                    }
-                                }
-                                _ => s.push(escaped_c), // handles \", \', and \\ naturally
-                            }
-                        }
-                    } else if n == quote {
-                        chars.next();
-                        break;
-                    } else {
-                        s.push(n);
-                        chars.next();
-                    }
-                }
-                tokens.push(ExprToken::Val(ConditionValue::Literal(s)));
-            }
-            _ => {
-                let mut ident = String::new();
-                while let Some(&(_, n)) = chars.peek() {
-                    if n.is_whitespace() || "()=!<>&|\"'".contains(n) {
-                        break;
-                    }
-                    ident.push(n);
-                    chars.next();
-                }
-                match ident.to_lowercase().as_str() {
-                    "and" => tokens.push(ExprToken::And),
-                    "or" => tokens.push(ExprToken::Or),
-                    "not" => tokens.push(ExprToken::Not),
-                    "true" | "false" => tokens.push(ExprToken::Val(ConditionValue::Literal(ident))),
-                    _ => {
-                        if let Ok(num) = ident.parse::<f64>() {
-                            tokens.push(ExprToken::Val(ConditionValue::Number(num)));
-                        } else if let Some(var) = ident.strip_prefix('$') {
-                            tokens
-                                .push(ExprToken::Val(ConditionValue::Variable(var.to_lowercase())));
-                        } else if let Some((ent, prop)) = ident.rsplit_once('.') {
-                            tokens.push(ExprToken::Val(ConditionValue::EntityProperty(
-                                ent.to_lowercase(),
-                                prop.to_string(),
-                            )));
-                        } else {
-                            tokens.push(ExprToken::Val(ConditionValue::Literal(ident)));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    Ok(tokens)
-}
-
-struct ExprParser<'a> {
-    tokens: &'a [ExprToken],
-    pos: usize,
-    depth: usize,
-    max_depth: usize,
-}
-
-impl<'a> ExprParser<'a> {
-    fn parse(tokens: &'a [ExprToken], max_depth: usize) -> Result<Condition, String> {
-        let mut parser = Self {
-            tokens,
-            pos: 0,
-            depth: 0,
-            max_depth,
-        };
-        let expr = parser.parse_or()?;
-        if parser.pos < parser.tokens.len() {
-            return Err("Unexpected trailing tokens in expression".into());
-        }
-        Ok(expr)
-    }
-    fn peek(&self) -> Option<&ExprToken> {
-        self.tokens.get(self.pos)
-    }
-    fn consume(&mut self) -> Option<&ExprToken> {
-        let t = self.tokens.get(self.pos);
-        self.pos += 1;
-        t
-    }
-
-    fn parse_or(&mut self) -> Result<Condition, String> {
-        self.depth += 1;
-        if self.depth > self.max_depth {
-            return Err(format!(
-                "Maximum expression nesting depth of {} exceeded",
-                self.max_depth
-            ));
-        }
-        let mut left = self.parse_and()?;
-        while let Some(ExprToken::Or) = self.peek() {
-            self.consume();
-            left = Condition::Or(Box::new(left), Box::new(self.parse_and()?));
-        }
-        self.depth -= 1;
-        Ok(left)
-    }
-    fn parse_and(&mut self) -> Result<Condition, String> {
-        self.depth += 1;
-        if self.depth > self.max_depth {
-            return Err(format!(
-                "Maximum expression nesting depth of {} exceeded",
-                self.max_depth
-            ));
-        }
-        let mut left = self.parse_comparison()?;
-        while let Some(ExprToken::And) = self.peek() {
-            self.consume();
-            left = Condition::And(Box::new(left), Box::new(self.parse_comparison()?));
-        }
-        self.depth -= 1;
-        Ok(left)
-    }
-    fn parse_comparison(&mut self) -> Result<Condition, String> {
-        self.depth += 1;
-        if self.depth > self.max_depth {
-            return Err(format!(
-                "Maximum expression nesting depth of {} exceeded",
-                self.max_depth
-            ));
-        }
-        let left_expr = self.parse_unary()?;
-        let res = match self.peek() {
-            Some(
-                op @ (ExprToken::Eq
-                | ExprToken::NotEq
-                | ExprToken::Lt
-                | ExprToken::Gt
-                | ExprToken::LtEq
-                | ExprToken::GtEq),
-            ) => {
-                let op_clone = op.clone();
-                self.consume();
-                let Condition::Value(left_val) = left_expr else {
-                    return Err("Left side of comparison must be a value".into());
-                };
-                let Condition::Value(right_val) = self.parse_unary()? else {
-                    return Err("Right side of comparison must be a value".into());
-                };
-                Ok(match op_clone {
-                    ExprToken::Eq => Condition::Eq(left_val, right_val),
-                    ExprToken::NotEq => Condition::NotEq(left_val, right_val),
-                    ExprToken::Lt => Condition::Lt(left_val, right_val),
-                    ExprToken::Gt => Condition::Gt(left_val, right_val),
-                    ExprToken::LtEq => Condition::LtEq(left_val, right_val),
-                    ExprToken::GtEq => Condition::GtEq(left_val, right_val),
-                    _ => return Err("Invalid comparison operator".into()),
-                })
-            }
-            _ => Ok(left_expr),
-        };
-        self.depth -= 1;
-        res
-    }
-    fn parse_unary(&mut self) -> Result<Condition, String> {
-        self.depth += 1;
-        if self.depth > self.max_depth {
-            return Err(format!(
-                "Maximum expression nesting depth of {} exceeded",
-                self.max_depth
-            ));
-        }
-        let res = if let Some(ExprToken::Not) = self.peek() {
-            self.consume();
-            Ok(Condition::Not(Box::new(self.parse_unary()?)))
-        } else {
-            match self.consume() {
-                Some(ExprToken::Val(v)) => Ok(Condition::Value(v.clone())),
-                Some(ExprToken::LParen) => {
-                    let expr = self.parse_or()?;
-                    if matches!(self.consume(), Some(ExprToken::RParen)) {
-                        Ok(expr)
-                    } else {
-                        Err("Expected ')'".into())
-                    }
-                }
-                _ => Err("Expected expression or value".into()),
-            }
-        };
-        self.depth -= 1;
-        res
     }
 }
 
@@ -1279,17 +921,9 @@ impl Template {
 
     fn parse_variable(content: &str) -> Result<Token, String> {
         let is_all_caps = is_all_caps(content);
-        let clean = content
-            .trim_start()
-            .strip_prefix('$')
-            .unwrap_or(content)
-            .trim();
+        let clean = content.trim_start().strip_prefix('$').unwrap_or(content);
 
-        let (key, fallback) = if let Some((k, f)) = clean.split_once("??") {
-            (k.trim(), Some(unescape_string(strip_outer_quotes(f))))
-        } else {
-            (clean, None)
-        };
+        let (key, fallback) = extract_variable_fallback(clean);
 
         reject_if(
             key.is_empty(),
@@ -1352,14 +986,8 @@ impl Template {
         };
 
         let (dynamic_key, dynamic_fallback) = if let Some(var_name) = base_verb.strip_prefix('$') {
-            if let Some((k, f)) = var_name.split_once("??") {
-                (
-                    Some(k.trim().to_string()),
-                    Some(unescape_string(strip_outer_quotes(f))),
-                )
-            } else {
-                (Some(var_name.trim().to_string()), None)
-            }
+            let (k, f) = extract_variable_fallback(var_name);
+            (Some(k.to_string()), f)
         } else {
             (None, None)
         };
@@ -1583,372 +1211,5 @@ impl TagFlags {
     #[must_use]
     pub const fn drop_possessive(self) -> bool {
         self.contains(Self::DROP_POSSESSIVE)
-    }
-}
-
-#[cold]
-pub(crate) fn validation_error(message: &str, content: &str, open_char: char) -> String {
-    let close_char = if open_char == TAG_ENTITY_OPEN {
-        TAG_ENTITY_CLOSE
-    } else {
-        TAG_VERB_CLOSE
-    };
-    let formatted_message = format!("{message}: {open_char}{content}{close_char}");
-    tracing::error!("{}", formatted_message);
-    formatted_message
-}
-
-#[inline]
-pub(crate) fn is_p_type(s: &str) -> bool {
-    let (clean, _) = parse_stance_prefixes(s);
-    let lower = clean.trim().to_lowercase();
-    matches!(
-        lower.as_str(),
-        "subj" | "obj" | "poss" | "abs_poss" | "reflex"
-    )
-}
-
-#[inline]
-pub(crate) fn consume_until_closed(
-    chars: &mut std::iter::Peekable<std::str::CharIndices<'_>>,
-    start_idx: usize,
-    close_char: char,
-    tag_type: &str,
-) -> Result<usize, String> {
-    let mut end_idx = start_idx + 1;
-    let mut closed = false;
-    let mut escaped = false;
-    let mut in_quote: Option<char> = None;
-    while let Some(&(j, ch)) = chars.peek() {
-        chars.next();
-        if escaped {
-            escaped = false;
-            continue;
-        }
-        if ch == '\\' {
-            escaped = true;
-            continue;
-        }
-        if let Some(q) = in_quote {
-            if ch == q {
-                in_quote = None;
-            }
-            continue;
-        }
-        if matches!(ch, '"' | '`') {
-            in_quote = Some(ch);
-            continue;
-        }
-        if ch == close_char {
-            end_idx = j;
-            closed = true;
-            break;
-        }
-    }
-
-    if !closed {
-        tracing::error!("Unclosed {} starting at index {}", tag_type, start_idx);
-        return Err(format!("Unclosed {tag_type} starting at index {start_idx}"));
-    }
-
-    Ok(end_idx)
-}
-
-/// Parses prefix modifiers (like `+`) used to force perspectives, returning the
-/// stripped string alongside the boolean flags for `force_3rd_person`/`force_article`, `no_smart`, and `force_singular`.
-#[inline]
-pub(crate) fn parse_stance_prefixes(mut s: &str) -> (&str, TagFlags) {
-    let mut flags = TagFlags::empty();
-    loop {
-        s = s.trim_start();
-        if let Some(stripped) = s.strip_prefix(MOD_FORCE_3RD_PERSON) {
-            flags.insert(TagFlags::FORCE_3RD_PERSON);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_NO_SMART) {
-            flags.insert(TagFlags::NO_SMART);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_FORCE_SINGULAR) {
-            flags.insert(TagFlags::FORCE_SINGULAR);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_PREFER_NOUN) {
-            flags.insert(TagFlags::PREFER_NOUN);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_ALLOW_AMBIGUOUS_YOU) {
-            flags.insert(TagFlags::ALLOW_AMBIGUOUS_YOU);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_EXTRACT_GROUP_MEMBER) {
-            flags.insert(TagFlags::EXTRACT_GROUP_MEMBER);
-            s = stripped;
-        } else if let Some(stripped) = s.strip_prefix(MOD_DROP_POSSESSIVE) {
-            flags.insert(TagFlags::DROP_POSSESSIVE);
-            s = stripped;
-        } else {
-            break;
-        }
-    }
-    (s, flags)
-}
-
-pub(crate) type ParsedForcedConjugations = (Option<Vec<String>>, Option<Vec<String>>);
-
-#[inline]
-pub(crate) fn parse_forced_conjugations(
-    forced: &str,
-    content: &str,
-) -> Result<ParsedForcedConjugations, String> {
-    let mut forced_present = None;
-    let mut forced_past = None;
-
-    let (pres_str, past_str) =
-        if let Some((present_overrides, past_overrides)) = forced.split_once(VERB_TENSE_SEP) {
-            (present_overrides, Some(past_overrides))
-        } else {
-            (forced, None)
-        };
-
-    if !pres_str.is_empty() {
-        let parts: Vec<String> = pres_str
-            .split(VERB_FORM_SEP)
-            .map(|s| s.trim().to_string())
-            .collect();
-        for part in &parts {
-            reject_if(
-                part.is_empty(),
-                "Verb tag has an empty forced present conjugation segment",
-                content,
-                TAG_VERB_OPEN,
-            )?;
-        }
-        reject_if(
-            parts.len() > 3,
-            "Verb tag has too many forced present conjugation segments",
-            content,
-            TAG_VERB_OPEN,
-        )?;
-        forced_present = Some(parts);
-    }
-
-    if let Some(past_overrides_str) = past_str {
-        reject_if(
-            past_overrides_str.is_empty(),
-            "Verb tag has an empty forced past conjugation segment",
-            content,
-            TAG_VERB_OPEN,
-        )?;
-        let parts: Vec<String> = past_overrides_str
-            .split(VERB_FORM_SEP)
-            .map(|s| s.trim().to_string())
-            .collect();
-        for part in &parts {
-            reject_if(
-                part.is_empty(),
-                "Verb tag has an empty forced past conjugation segment",
-                content,
-                TAG_VERB_OPEN,
-            )?;
-        }
-        reject_if(
-            parts.len() > 3,
-            "Verb tag has too many forced past conjugation segments",
-            content,
-            TAG_VERB_OPEN,
-        )?;
-        forced_past = Some(parts);
-    }
-
-    Ok((forced_present, forced_past))
-}
-
-#[inline]
-pub(crate) fn split_tag<'a>(
-    content: &'a str,
-    open_char: char,
-    malformed_msg: &str,
-) -> Result<(&'a str, Option<&'a str>), String> {
-    let mut parts = content.split(TAG_SEPARATOR);
-    let p1 = parts.next().unwrap_or_default().trim();
-    let p2 = parts.next().map(str::trim);
-    reject_if(parts.next().is_some(), malformed_msg, content, open_char)?;
-    Ok((p1, p2))
-}
-
-#[inline]
-pub(crate) fn push_literal(tokens: &mut Vec<Token>, raw: &str, start: usize, end: usize) {
-    if end > start
-        && let Some(slice) = raw.get(start..end)
-    {
-        if let Some(Token::Literal(last)) = tokens.last_mut() {
-            last.push_str(slice);
-        } else {
-            tokens.push(Token::Literal(slice.to_string()));
-        }
-    }
-}
-
-#[inline]
-pub(crate) fn unescape_string(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let mut chars = input.chars().peekable();
-    while let Some(c) = chars.next() {
-        if c == '\\' {
-            if let Some(escaped) = chars.next() {
-                match escaped {
-                    'n' => out.push('\n'),
-                    'r' => out.push('\r'),
-                    't' => out.push('\t'),
-                    'u' => {
-                        if chars.peek() == Some(&'{') {
-                            chars.next(); // Consume '{'
-                            let mut hex = String::new();
-                            while let Some(&hc) = chars.peek() {
-                                chars.next();
-                                if hc == '}' {
-                                    break;
-                                }
-                                hex.push(hc);
-                            }
-                            if let Ok(code) = u32::from_str_radix(&hex, 16) {
-                                out.push(
-                                    char::from_u32(code).unwrap_or(char::REPLACEMENT_CHARACTER),
-                                );
-                            } else {
-                                out.push(char::REPLACEMENT_CHARACTER);
-                            }
-                        } else {
-                            out.push('u');
-                        }
-                    }
-                    _ => out.push(escaped),
-                }
-            }
-        } else {
-            out.push(c);
-        }
-    }
-    out
-}
-
-/// Safely strips exactly one layer of outer quotes if they match.
-#[inline]
-pub(crate) fn strip_outer_quotes(s: &str) -> &str {
-    let s = s.trim();
-    if let Some(first) = s.chars().next()
-        && matches!(first, '"' | '\'' | '`')
-        && s.ends_with(first)
-        && s.len() >= 2
-    {
-        return &s[1..s.len() - 1];
-    }
-    s
-}
-
-#[inline]
-pub(crate) fn is_article(s: &str) -> bool {
-    const ARTICLES: &[&str] = &[
-        "a",
-        "an",
-        "the",
-        "this",
-        "that",
-        "another",
-        "one",
-        "one of",
-        "one of the",
-        "some",
-    ];
-    ARTICLES.iter().any(|&art| s.eq_ignore_ascii_case(art))
-}
-
-#[inline]
-pub(crate) fn is_indefinite_article(s: &str) -> bool {
-    s.eq_ignore_ascii_case("a") || s.eq_ignore_ascii_case("an")
-}
-
-#[inline]
-pub(crate) const fn is_escapable_char(c: char) -> bool {
-    matches!(
-        c,
-        TAG_ENTITY_OPEN | TAG_VERB_OPEN | TAG_ENTITY_CLOSE | TAG_VERB_CLOSE | TAG_ESCAPE
-    )
-}
-
-#[inline]
-pub(crate) fn is_capitalized(s: &str) -> bool {
-    s.chars().next().is_some_and(char::is_uppercase)
-}
-
-#[inline]
-pub(crate) fn is_all_caps(s: &str) -> bool {
-    let has_letters = s.chars().any(char::is_alphabetic);
-    has_letters
-        && s.chars()
-            .filter(|c| c.is_alphabetic())
-            .all(char::is_uppercase)
-}
-
-#[inline]
-pub(crate) fn reject_if(
-    condition: bool,
-    error_msg: &str,
-    content: &str,
-    open_char: char,
-) -> Result<(), String> {
-    if condition {
-        Err(validation_error(error_msg, content, open_char))
-    } else {
-        Ok(())
-    }
-}
-
-#[inline]
-pub(crate) fn validate_property_segments(
-    path: &str,
-    error_msg: &str,
-    content: &str,
-    open_char: char,
-) -> Result<(), String> {
-    reject_if(
-        path.split(TAG_PROPERTY_SEP).any(str::is_empty),
-        error_msg,
-        content,
-        open_char,
-    )
-}
-
-/// Parses possessive suffix `'s`, returning the stripped string and a boolean flag.
-#[inline]
-pub(crate) fn parse_possessive_suffix(s: &str) -> (&str, bool) {
-    if let Some(stripped) = s.strip_suffix(MOD_POSSESSIVE) {
-        (stripped, true)
-    } else if let Some(stripped) = s.strip_suffix("'S") {
-        (stripped, true)
-    } else if let Some(stripped) = s.strip_suffix("’s") {
-        (stripped, true)
-    } else if let Some(stripped) = s.strip_suffix("’S") {
-        (stripped, true)
-    } else if let Some(stripped) = s.strip_suffix('\'') {
-        (stripped, true)
-    } else if let Some(stripped) = s.strip_suffix('’') {
-        (stripped, true)
-    } else {
-        (s, false)
-    }
-}
-
-/// Finds a possessive suffix with a trailing space, returning the index and the byte length of the match.
-#[inline]
-pub(crate) fn find_spaced_possessive(s: &str) -> Option<(usize, usize)> {
-    if let Some(idx) = s.find("'s ") {
-        Some((idx, 3))
-    } else if let Some(idx) = s.find("'S ") {
-        Some((idx, 3))
-    } else if let Some(idx) = s.find("’s ") {
-        Some((idx, 5))
-    } else if let Some(idx) = s.find("’S ") {
-        Some((idx, 5))
-    } else if let Some(idx) = s.find("' ") {
-        Some((idx, 2))
-    } else {
-        s.find("’ ").map(|idx| (idx, 4))
     }
 }
